secured crud endpoints

+++ src/portfolio/server/Controllers/StoryController.cs
@@ -9,13 +9,16 @@ namespace server.Controllers;
 [Route("/api")]
 public class StoryController : ControllerBase
 {
-    private readonly ILogger<StoryController> _logger;
+    private const string SESSION_KEY_NAME = "_SessionId";
+    private readonly ILogger<StoryController> logger;
     private readonly IStoryDbService dbService;
+    private readonly ISessionService sessionService;
 
-    public StoryController(ILogger<StoryController> logger, IStoryDbService dbService)
+    public StoryController(ILogger<StoryController> logger, IStoryDbService dbService, ISessionService sessionService)
     {
-        _logger = logger;
+        this.logger = logger;
         this.dbService = dbService;
+        this.sessionService = sessionService;
     }
 
     [HttpGet]
@@ -29,29 +32,77 @@ public class StoryController : ControllerBase
     [Route("stories/{id:int}")]
     public async Task<Story?> GetStoryById(int id)
     {
-        _logger.LogDebug("Getting story with id {id}", id);
+        logger.LogDebug("Getting story with id {id}", id);
         return await dbService.GetStoryByIdAsync(id);
     }
 
     [HttpPost]
     [Route("stories")]
-    public async Task<Story> PostStory(Story story)
+    public async Task<ActionResult<Story>> PostStory(Story story)
     {
-        story.Id = null;
-        return await dbService.AddStoryAsync(story);
+        if (HttpContext.Request.Cookies.TryGetValue(SESSION_KEY_NAME, out var sessionId))
+        {
+            try
+            {
+                if (await sessionService.IsSessionValidAsync(sessionId))
+                {
+                    story.Id = null;
+                    return await dbService.AddStoryAsync(story);
+                }
+            }
+            catch (Exception ex)
+            {
+                logger.LogError("Could not post story. Ex {ex}", ex);
+                return StatusCode(500);
+            }
+        }
+
+        return StatusCode(403);
     }
 
     [HttpPut]
     [Route("stories/{id:int}")]
-    public async Task<Story> UpdateStory(Story story, int id)
+    public async Task<ActionResult<Story>> UpdateStory(Story story, int id)
     {
-        return await dbService.UpdateStoryAsync(id, story);
+        if (HttpContext.Request.Cookies.TryGetValue(SESSION_KEY_NAME, out var sessionId))
+        {
+            try
+            {
+                if (await sessionService.IsSessionValidAsync(sessionId))
+                {
+                    return await dbService.UpdateStoryAsync(id, story);
+                }
+            }
+            catch (Exception ex)
+            {
+                logger.LogError("Could not update story. Ex {ex}", ex);
+                return StatusCode(500);
+            }
+        }
+
+        return StatusCode(403);
     }
 
     [HttpDelete]
     [Route("stories/{id:int}")]
-    public async Task<Story> DeleteStory(int id)
+    public async Task<ActionResult<Story>> DeleteStory(int id)
     {
-        return await dbService.DeleteStoryAsync(id);
+        if (HttpContext.Request.Cookies.TryGetValue(SESSION_KEY_NAME, out var sessionId))
+        {
+            try
+            {
+                if (await sessionService.IsSessionValidAsync(sessionId))
+                {
+                    return await dbService.DeleteStoryAsync(id);
+                }
+            }
+            catch (Exception ex)
+            {
+                logger.LogError("Could not delete story. Ex {ex}", ex);
+                return StatusCode(500);
+            }
+        }
+
+        return StatusCode(403);
     }
 }
